Simple fix for Postgres code to get it to compile on Ubuntu 16.04 / g++ 5.3.1 -- I dedicate this commit to the memory of @jarulraj

+++ src/postgres/backend/access/gin/ginlogic.cpp
@@ -168,7 +168,7 @@ shimTriConsistentFn(GinScanKey key)
 	nmaybe = 0;
 	for (i = 0; i < key->nentries; i++)
 	{
-		if (key->entryRes[i] == GIN_MAYBE)
+		if (static_cast<int>(key->entryRes[i]) == GIN_MAYBE)
 		{
 			if (nmaybe >= MAX_MAYBE_ENTRIES)
 				return GIN_MAYBE;
